+++ packages/plugin-hardhat/src/utils/options.ts
@@ -20,5 +20,8 @@ export interface DeployProxyOptions extends Options {
 
 export interface UpgradeProxyOptions extends Options {
   call?: { fn: string; args?: unknown[] } | string;
-  unsafeSkipStorageCheck?: boolean;
+}
+
+export interface PrepareUpgradeOptions extends Options {
+  getTxResponse?: boolean;
 }